+++ models/player_current_inventory.go
@@ -4,9 +4,9 @@ import "gorm.io/gorm"
 
 type PlayerCurrentInventory struct {
 	gorm.Model
-	PlayerId   int64
+	PlayerId   uint
 	Player     Player
-	ItemId     int64
+	ItemId     uint
 	Item       Item
 	IsBound    bool `gorm:"default:false"`
 	IsEquipped bool `gorm:"default:false"`
